+++ packages/txservice/src/dispatch.ts
@@ -11,7 +11,6 @@ import {
 } from "@connext/nxtp-utils";
 import interval from "interval-promise";
 
-import { Gas, WriteTransaction, OnchainTransaction, TransactionBuffer } from "./shared";
 import {
   BadNonce,
   TransactionReplaced,
@@ -22,8 +21,12 @@ import {
   TransactionProcessingError,
   NotEnoughConfirmations,
   TransactionAlreadyKnown,
-} from "./shared/error";
-import { ChainConfig, TransactionServiceConfig } from "./config";
+  Gas,
+  WriteTransaction,
+  OnchainTransaction,
+  TransactionBuffer,
+} from "./shared";
+import { ChainConfig } from "./config";
 import { ChainRpcProvider } from "./provider";
 
 export type DispatchCallbacks = {
@@ -75,13 +78,12 @@ export class TransactionDispatch extends ChainRpcProvider {
   constructor(
     logger: Logger,
     public readonly chainId: number,
-    chainConfig: ChainConfig,
-    config: TransactionServiceConfig,
+    config: ChainConfig,
     signer: string | Signer,
     private readonly callbacks: DispatchCallbacks,
     startLoops = true,
   ) {
-    super(logger, chainId, chainConfig, config, signer);
+    super(logger, chainId, config, signer);
     this.inflightBuffer = new TransactionBuffer(logger, TransactionDispatch.MAX_INFLIGHT_TRANSACTIONS, {
       name: "INFLIGHT",
       chainId: this.chainId,
@@ -379,8 +381,8 @@ export class TransactionDispatch extends ChainRpcProvider {
               nonce,
               gas,
               {
-                confirmationTimeout: this.confirmationTimeout,
-                confirmationsRequired: this.confirmationsRequired,
+                confirmationTimeout: this.config.confirmationTimeout,
+                confirmationsRequired: this.config.confirmations,
               },
               txsId,
             );
@@ -664,10 +666,10 @@ export class TransactionDispatch extends ChainRpcProvider {
 
     // Here we wait for the target confirmations.
     // TODO: Ensure we are comfortable with how this timeout period is calculated.
-    const timeout = this.confirmationTimeout * this.confirmationsRequired * 2;
+    const timeout = this.config.confirmationTimeout * this.config.confirmations * 2;
     let receipt: providers.TransactionReceipt;
     try {
-      receipt = await this.confirmTransaction(transaction, this.confirmationsRequired, timeout);
+      receipt = await this.confirmTransaction(transaction, this.config.confirmations, timeout);
     } catch (error) {
       this.logger.error(
         "Did not get enough confirmations for a *mined* transaction! Did a re-org occur?",
@@ -678,12 +680,12 @@ export class TransactionDispatch extends ChainRpcProvider {
           chainId: this.chainId,
           transaction: transaction.loggable,
           confirmations: transaction.receipt.confirmations,
-          confirmationsRequired: this.confirmationsRequired,
+          confirmationsRequired: this.config.confirmations,
         },
       );
       // No other errors should normally occur during this confirmation attempt. This could occur during a reorg.
       throw new NotEnoughConfirmations(
-        this.confirmationsRequired,
+        this.config.confirmations,
         transaction.receipt.transactionHash,
         transaction.receipt.confirmations,
         {
@@ -731,7 +733,7 @@ export class TransactionDispatch extends ChainRpcProvider {
     const { requestContext, methodContext } = createLoggingContext(this.bump.name, transaction.context);
     if (
       transaction.bumps >= transaction.hashes.length ||
-      transaction.gas.price.gte(BigNumber.from(this.config.gasMaximum))
+      transaction.gas.price.gte(BigNumber.from(this.config.gasPriceMaximum))
     ) {
       // If we've already bumped this tx but it's failed to resubmit, we should return here without bumping.
       // The number of gas bumps we've done should always be less than the number of txs we've submitted.
@@ -739,7 +741,7 @@ export class TransactionDispatch extends ChainRpcProvider {
         chainId: this.chainId,
         bumps: transaction.bumps,
         gasPrice: utils.formatUnits(transaction.gas.price, "gwei"),
-        gasMaximum: utils.formatUnits(this.config.gasMaximum, "gwei"),
+        gasMaximum: utils.formatUnits(this.config.gasPriceMaximum, "gwei"),
       });
       return;
     }
@@ -748,14 +750,14 @@ export class TransactionDispatch extends ChainRpcProvider {
     // Get the current gas baseline price, in case it's changed drastically in the last block.
     let updatedPrice: BigNumber;
     try {
-      updatedPrice = await this.getGasPrice(requestContext);
+      updatedPrice = await this.getGasPrice(requestContext, false);
     } catch {
-      updatedPrice = BigNumber.from(this.config.gasMinimum);
+      updatedPrice = BigNumber.from(this.config.gasPriceMinimum);
     }
     const determinedBaseline = updatedPrice.gt(previousPrice) ? updatedPrice : previousPrice;
     // Scale up gas by percentage as specified by config.
     transaction.gas.price = determinedBaseline
-      .add(determinedBaseline.mul(this.config.gasReplacementBumpPercent).div(100))
+      .add(determinedBaseline.mul(this.config.gasPriceReplacementBumpPercent).div(100))
       .add(1);
     this.logger.info(`Tx bumped.`, requestContext, methodContext, {
       chainId: this.chainId,